pyproject.toml
[project]
name = "nbhelper-onebot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "冰犬🧊🧊🧊", email = "hf4minecraft@163.com"},
]
dependencies = ["nonebot2[websockets]>=2.4.2", "nonebot-adapter-onebot>=2.4.6", "cachetools>=6.0.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
